chore: 일부 코드 리팩토링
