+++ src/pages/schedule/monthly_schedule/MonthlyScheduleView.tsx
@@ -1,18 +1,12 @@
 import Calendar from "react-calendar";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "./Calendar.scss";
-import EventTagBox from "./EventTagBox";
-import { dateFormatter } from "@/utils/dateFormatter";
 import type { UserScheduleData } from "@/apis/participate/participateTypes";
 import { useMonthlySwiperController } from "@/hooks/useMonthlySwiperController";
+import CalendarTile from "./MonthlyCalendarTile";
 
 interface Props {
   schedules: UserScheduleData[];
-}
-
-export interface Event {
-  id: string;
-  name: string;
 }
 
 const MonthlyScheduleView = ({ schedules }: Props) => {
@@ -40,19 +34,7 @@ const MonthlyScheduleView = ({ schedules }: Props) => {
                 view="month"
                 onViewChange={() => {}}
                 showNeighboringMonth={false}
-                tileContent={({ date }) => {
-                  const eventNames: Event[] = [];
-                  if (schedules) {
-                    schedules.forEach((schedule) => {
-                      const date1 = dateFormatter(new Date(schedule.startAt));
-                      const date2 = dateFormatter(new Date(date));
-                      if (date1 === date2) {
-                        eventNames.push({ id: schedule.eventId, name: schedule.title });
-                      }
-                    });
-                  }
-                  return <EventTagBox eventNames={eventNames} />;
-                }}
+                tileContent={({ date }) => <CalendarTile date={date} schedules={schedules} />}
                 formatDay={(_, date) => date.toLocaleString("en", { day: "numeric" })}
                 formatShortWeekday={(_, date) => date.toLocaleString("en-US", { weekday: "short" })}
                 onActiveStartDateChange={() => {}}
